+++ MasterPassword/C/src/mpw_core.h
@@ -1,12 +1,14 @@
 #ifndef _MPW_CORE_H_
 #define _MPW_CORE_H_
 
+#include <stdint.h>
+
 int mpw_core(char * const password, const size_t passLen, char const * const userName,
 	char const * const masterPassword, char const * const siteTypeString, char const * const siteName,
 	const int siteCounter);
 
 int mpw_core_calculate_master_key_salt(char const * const mpNameSpace, char const * const userName,
-                                       char * const masterKeySalt, unsigned long * const masterKeySaltLength );
+                                       char * const masterKeySalt, size_t * const masterKeySaltLength );
 
 int mpw_core_calculate_site_seed( char * const sitePasswordInfo, size_t * const sitePasswordInfoLength,
                                  char const * const mpNameSpace, char const * const siteName, int siteCounter );
